feat: restructure local_tools
