+++ composio/local_tools/file/__init__.py
@@ -1,3 +1,4 @@
 """
 File manager.
 """
+from .tool import FileTool
