+++ composio/local_tools/file/tool.py
@@ -4,9 +4,8 @@ File I/O tool for Composio.
 
 import typing as t
 
-from ..tool import Tool, Action
-from .actions.read_file import ReadFile
-from .actions.write_file import WriteFile
+from .actions import ReadFile, WriteFile
+from composio.tools.local import Tool, Action
 
 class FileTool(Tool):
     """File I/O tool."""
