+++ composio/local_tools/greptile/__init__.py
@@ -1,2 +1,1 @@
-from .actions.codequery import CodeQuery
 from .tool import Greptile
