+++ composio/local_tools/greptile/actions/codequery.py
@@ -1,16 +1,12 @@
 import json
 import os
-
 import requests
 from pydantic import BaseModel, Field
 
-from ...tool import Action
+from composio.tools.local import Action
 from composio.local_tools.local_workspace.commons.get_logger import get_logger
 
-
-# Greptile query tool. Given the github repo and the question, it will return the answer to the question.
 logger = get_logger()
-
 
 class message(BaseModel):
     id: str = Field(..., description="The id of the message")
@@ -20,7 +16,6 @@ class message(BaseModel):
         description="The role of the message. ex system,user",
         examples=["system", "user"],
     )
-
 
 class CodeQueryRequest(BaseModel):
     question: str = Field(
@@ -42,7 +37,6 @@ class CodeQueryRequest(BaseModel):
         description="The repository to ask the question about. This should be a github repository. Example openai/docs, samparkai/composio",
         examples=["openai/docs", "samparkai/composio"],
     )
-
 
 class CodeQueryResponse(BaseModel):
     response: str = Field(..., description="The response to the question")
