--- composio/local_tools/local_workspace/Readme.md
@@ -1,11 +1,0 @@
-1. set your github token in environment 
-    GITHUB_ACCESS_TOKEN=""
-2. set a github repo, that you want to solve issue from 
-    HARD_CODED_REPO_NAME=""
-3. Install Docker, then start Docker locally.
-4. Run `docker pull sweagent/swe-agent:latest`
-5. use your tools repo venv 
-6. `cd ~/hermes/tools/services/swelib/local_workspace`
-7. run `python run_with_crewai_local_tool.py`
-8. change the issue according to your task / or anything you want to fix -- give it a try 🚀🚀
-9. task description is given in file `cat ~/hermes/tools/services/swelib/local_workspace/agent_task_data.yaml`
