+++ composio/local_tools/local_workspace/cmd_manager/actions/base_class.py
@@ -4,9 +4,9 @@ from typing import Optional
 
 from pydantic import BaseModel, Field
 
-from composio.local_tools.action import Action
-from composio.local_tools.local_workspace.commons.get_logger import get_logger
-from composio.local_tools.local_workspace.commons.history_processor import (
+from composio.tools.local import Action
+from composio.local_tools.local_workspace.commons import (
+    get_logger,
     HistoryProcessor,
 )
 from composio.local_tools.local_workspace.commons.local_docker_workspace import (
