+++ composio/local_tools/local_workspace/cmd_manager/cmd_manager_tool.py
@@ -1,32 +1,23 @@
 from typing import Optional
+from composio.tools.local import Tool
 
-from composio.local_tools.local_workspace.cmd_manager.actions.clone_github import (
+from composio.local_tools.local_workspace.cmd_manager.actions import (
     GithubCloneCmd,
-)
-from composio.local_tools.local_workspace.cmd_manager.actions.cmds import (
     CreateFileCmd,
     GoToLineNumInOpenFile,
     OpenFile,
-)
-from composio.local_tools.local_workspace.cmd_manager.actions.edit_cmd import EditFile
-from composio.local_tools.local_workspace.cmd_manager.actions.run_cmd import (
+    Scroll,
     RunCommandOnWorkspace,
-)
-from composio.local_tools.local_workspace.cmd_manager.actions.scroll_cmds import Scroll
-from composio.local_tools.local_workspace.cmd_manager.actions.search_cmds import (
     FindFileCmd,
     GetCurrentDirCmd,
     SearchDirCmd,
     SearchFileCmd,
+    EditFile,
 )
-from composio.local_tools.local_workspace.commons.history_processor import (
+from composio.local_tools.local_workspace.commons import (
     HistoryProcessor,
+    WorkspaceManagerFactory
 )
-from composio.local_tools.local_workspace.commons.local_docker_workspace import (
-    WorkspaceManagerFactory,
-)
-from composio.local_tools.tool import Tool
-
 
 class CmdManagerTool(Tool):
     """
