+++ composio/local_tools/local_workspace/history_keeper/history_keeper_tool.py
@@ -1,20 +1,15 @@
+import typing as t
 from typing import Optional
+from composio.tools.local import Tool, Action
 
 from composio.local_tools.local_workspace.commons.get_logger import get_logger
-from composio.local_tools.local_workspace.commons.history_processor import (
-    HistoryProcessor,
-)
-from composio.local_tools.local_workspace.commons.local_docker_workspace import (
+from composio.local_tools.local_workspace.commons import (
     WorkspaceManagerFactory,
+    HistoryProcessor
 )
-from composio.local_tools.local_workspace.history_keeper.actions.get_workspace_history import (
-    GetWorkspaceHistory,
-)
-from composio.local_tools.tool import Tool
-
+from .actions import GetWorkspaceHistory
 
 logger = get_logger()
-
 
 class HistoryKeeper(Tool):
     """
@@ -24,7 +19,7 @@ class HistoryKeeper(Tool):
     workspace_factory: Optional[WorkspaceManagerFactory] = None
     history_processor: Optional[HistoryProcessor] = None
 
-    def actions(self) -> list:
+    def actions(self) -> list[t.Type[Action]]:
         return [GetWorkspaceHistory]
 
     def triggers(self) -> list:
