+++ composio/local_tools/mathematical/actions/calculator.py
@@ -1,7 +1,5 @@
 from pydantic import BaseModel, Field
-
-from ...tool import Action
-
+from composio.tools.local import Action
 
 class CalculatorRequest(BaseModel):
     operation: str = Field(
@@ -9,10 +7,8 @@ class CalculatorRequest(BaseModel):
         description="A mathematical expression, a couple examples are `200*7` or `5000/2*10`",
     )
 
-
 class CalculatorResponse(BaseModel):
     result: str = Field(..., description="Result of the calculation")
-
 
 class Calculator(Action):
     """
