+++ composio/local_tools/ragtool/__init__.py
@@ -1,1 +1,1 @@
-from .ragtool import RagToolActions
+from .tool import RagTool
