+++ composio/local_tools/webtool/__init__.py
@@ -1,1 +1,1 @@
-from .WebTool import WebTool
+from .tool import WebTool
