+++ composio/tools/local/tool.py
@@ -9,7 +9,6 @@ from composio.local_tools.local_workspace.commons.local_docker_workspace import 
 
 from .action import Action
 
-
 class Tool:
     @property
     def tool_name(self) -> str:
